Implemented RecyclerView with ListAdapter, added onLike and onShare listeners

+++ app/src/main/java/ru/netology/nmedia/Utils.kt
@@ -2,7 +2,7 @@ package ru.netology.nmedia
 
 import kotlin.math.floor
 
-fun formatCount(count: Int): String {
+fun formatCount(count: Long): String {
     return when {
         count < 1_000 -> count.toString()
         count < 10_000 -> {
